Extract validators into standalone file
